Simplify Header and ListHeader
