--- app/views/RoomsListView/components/Header/Header.tsx
@@ -1,88 +1,0 @@
-import React from 'react';
-import { StyleSheet, Text, View, useWindowDimensions } from 'react-native';
-import { TouchableOpacity } from 'react-native-gesture-handler';
-import { type TouchableOpacityProps } from 'react-native-gesture-handler';
-
-import I18n from '../../../../i18n';
-import sharedStyles from '../../../Styles';
-import { useTheme } from '../../../../theme';
-import SearchHeader from '../../../../containers/SearchHeader';
-import { useAppSelector } from '../../../../lib/hooks';
-
-const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		justifyContent: 'center'
-	},
-	button: {
-		flexDirection: 'row',
-		alignItems: 'center'
-	},
-	title: {
-		flexShrink: 1,
-		fontSize: 16,
-		...sharedStyles.textSemibold
-	},
-	subtitle: {
-		fontSize: 14,
-		...sharedStyles.textRegular
-	}
-});
-
-interface IRoomHeader {
-	connecting: boolean;
-	connected: boolean;
-	isFetching: boolean;
-	serverName: string;
-	server: string;
-	onPress: TouchableOpacityProps['onPress'];
-	search: (text: string) => void;
-	searchEnabled: boolean;
-}
-
-const Header = React.memo(
-	({ connecting, connected, isFetching, serverName = 'Rocket.Chat', server, onPress, search, searchEnabled }: IRoomHeader) => {
-		const { status: supportedVersionsStatus } = useAppSelector(state => state.supportedVersions);
-		const { colors } = useTheme();
-		const { fontScale } = useWindowDimensions();
-
-		if (searchEnabled) {
-			// This value is necessary to keep the alignment in MasterDetail.
-			const height = 37 * fontScale;
-			return <SearchHeader onSearchChangeText={search} testID='rooms-list-view-search-input' style={{ height }} />;
-		}
-		let subtitle;
-		if (supportedVersionsStatus === 'expired') {
-			subtitle = 'Cannot connect';
-		} else if (connecting) {
-			subtitle = I18n.t('Connecting');
-		} else if (isFetching) {
-			subtitle = I18n.t('Updating');
-		} else if (!connected) {
-			subtitle = I18n.t('Waiting_for_network');
-		} else {
-			subtitle = server?.replace(/(^\w+:|^)\/\//, '');
-		}
-		return (
-			<View style={styles.container} accessibilityLabel={`${serverName} ${subtitle}`} accessibilityRole='header' accessible>
-				<TouchableOpacity onPress={onPress} testID='rooms-list-header-servers-list-button'>
-					<View style={styles.button}>
-						<Text style={[styles.title, { color: colors.fontTitlesLabels }]} numberOfLines={1}>
-							{serverName}
-						</Text>
-					</View>
-					{subtitle ? (
-						<Text
-							testID='rooms-list-header-server-subtitle'
-							style={[styles.subtitle, { color: colors.fontSecondaryInfo }]}
-							numberOfLines={1}>
-							{subtitle}
-						</Text>
-					) : null}
-				</TouchableOpacity>
-			</View>
-		);
-	}
-);
-
-export default Header;
