+++ app/views/RoomsListView/index.tsx
@@ -1,32 +1,32 @@
-import React, { memo, useContext, useEffect } from 'react';
+import { useNavigation } from '@react-navigation/native';
+import React, { memo, useContext } from 'react';
 import { FlatList, RefreshControl } from 'react-native';
 import { useSafeAreaFrame } from 'react-native-safe-area-context';
-import { useNavigation } from '@react-navigation/native';
 import { shallowEqual } from 'react-redux';
 
-import RoomItem from '../../containers/RoomItem';
 import ActivityIndicator from '../../containers/ActivityIndicator';
-import { useTheme } from '../../theme';
-import { getUserSelector } from '../../selectors/login';
-import styles from './styles';
-import ListHeader from './components/ListHeader';
-import { getRoomAvatar, getRoomTitle, getUidDirectMessage, isRead, isIOS, isTablet } from '../../lib/methods/helpers';
-import { MAX_SIDEBAR_WIDTH } from '../../lib/constants';
-import Container from './components/Container';
-import { useSubscriptions } from './hooks/useSubscriptions';
-import { useAppSelector } from '../../lib/hooks/useAppSelector';
-import { useHeader } from './hooks/useHeader';
-import { events, logEvent } from '../../lib/methods/helpers/log';
-import { goRoom } from '../../lib/methods/helpers/goRoom';
-import { SectionHeader } from './components/SectionHeader';
-import { useGetItemLayout } from './hooks/useGetItemLayout';
-import { useRefresh } from './hooks/useRefresh';
-import { RoomsProvider, RoomsContext } from './RoomsSearchProvider';
-import { IRoomItem } from './definitions';
-import { SupportedVersionsExpired } from '../../containers/SupportedVersions';
-import { ChangePasswordRequired } from '../../containers/ChangePasswordRequired';
 import BackgroundContainer from '../../containers/BackgroundContainer';
+import { ChangePasswordRequired } from '../../containers/ChangePasswordRequired';
+import RoomItem from '../../containers/RoomItem';
+import { SupportedVersionsExpired } from '../../containers/SupportedVersions';
 import i18n from '../../i18n';
+import { MAX_SIDEBAR_WIDTH } from '../../lib/constants';
+import { useAppSelector } from '../../lib/hooks/useAppSelector';
+import { getRoomAvatar, getRoomTitle, getUidDirectMessage, isIOS, isRead, isTablet } from '../../lib/methods/helpers';
+import { goRoom } from '../../lib/methods/helpers/goRoom';
+import { events, logEvent } from '../../lib/methods/helpers/log';
+import { getUserSelector } from '../../selectors/login';
+import { useTheme } from '../../theme';
+import Container from './components/Container';
+import ListHeader from './components/ListHeader';
+import { SectionHeader } from './components/SectionHeader';
+import { IRoomItem } from './definitions';
+import { useGetItemLayout } from './hooks/useGetItemLayout';
+import { useHeader } from './hooks/useHeader';
+import { useRefresh } from './hooks/useRefresh';
+import { useSubscriptions } from './hooks/useSubscriptions';
+import { RoomsContext, RoomsProvider } from './RoomsSearchProvider';
+import styles from './styles';
 
 const INITIAL_NUM_TO_RENDER = isTablet ? 20 : 12;
